Add Files for Java Exercise
Implement a program that takes an integer count as an argument on the
command line.  This count is the number of random double precision
floating point numbers [0.0-1.0) to generate.  Count how many times the
digit in the first decimal place occurs in the sequence.  Print out each
digit and the number of times it occurred in ascending order of the
number of times it occurred